Api功能更新

+++ zhili-home-project/src/test/java/com/zhouzhili/zhilihomeproject/mockdata/UserMockData.java
@@ -9,6 +9,7 @@ import com.zhouzhili.zhilihomeproject.repository.security.RoleRepository;
 import com.zhouzhili.zhilihomeproject.repository.security.UserRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -37,9 +38,10 @@ public class UserMockData {
     private PasswordEncoder passwordEncoder;
 
     @Test
+    @DisplayName("添加角色信息")
     public void testRoleSave() {
         List<Role> roles = roleRepository.findAll();
-        if (roles.size() > 0) {
+        if (roles != null && roles.size() > 0) {
             roleRepository.deleteAll();
         }
         Role adminRole = new Role();
@@ -64,8 +66,11 @@ public class UserMockData {
         User admin = new User();
         Faker faker = new Faker();
         Name name = faker.name();
-        admin.setUsername(name.name());
+        admin.setUsername("admin");
         admin.setPassword(passwordEncoder.encode("admin"));
+        admin.setEmail(faker.bothify("????##@gmail.com"));
+        admin.setCreateDate(new Date());
+        admin.setUpdateDate(new Date());
         Optional<Role> adminRole = roleRepository.findRoleByRoleName("ROLE_ADMIN");
         Set<Role> roles = new HashSet<>();
         Assertions.assertNotNull(adminRole.get(), "未找到ROLE_ADMIN");
@@ -80,11 +85,15 @@ public class UserMockData {
         List<User> users = new ArrayList<>();
         Faker faker = new Faker();
         Name name = faker.name();
+
         List<Role> roles = roleRepository.findAll();
         for (int i = 0; i < 100; i++) {
             User user = new User();
             user.setUsername(name.username());
             user.setPassword(passwordEncoder.encode("123456"));
+            user.setCreateDate(new Date());
+            user.setUpdateDate(new Date());
+            user.setEmail(faker.bothify("????##@gmail.com"));
             Set<Role> roleSet = Set.of(roles.get((int) (Math.random() * 3)));
             user.setRoles(roleSet);
             users.add(user);
